Renamed wdparams to reflect new default
